* Parallel/Messages.h (SeqDataResponse::m_multiplicity): Change type to uint32_t. It must accomodate the coverage from both strands, which is 17 bits (2 times 16 bits).
git-svn-id: https://svn.bcgsc.ca/svn/abyss/trunk@54079 ad02020f-2a26-4133-947e-1829317c1416

+++ Parallel/Messages.h
@@ -170,7 +170,7 @@ class SeqDataResponse : public Message
 		IDType m_group;
 		IDType m_id;
 		ExtensionRecord m_extRecord;
-		uint16_t m_multiplicity;
+		uint32_t m_multiplicity;
 };
 
 class SetBaseMessage : public Message
